All endpoints in UsersController now working. Add empty methods to RolesController.

+++ relay-writer-api/src/main/java/se/melindasw/relaywriter/users/UsersServiceImplementation.java
@@ -1,6 +1,7 @@
 package se.melindasw.relaywriter.users;
 
 import org.springframework.stereotype.Service;
+import se.melindasw.relaywriter.exceptions.UserNotFoundException;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -37,40 +38,63 @@ public class UsersServiceImplementation implements UsersService {
 
   @Override
   public String deleteUser(Long userID) {
-    usersRepo.deleteById(userID);
-    return "User with id " + userID + " deleted.";
+    if (usersRepo.findById(userID).isPresent()) {
+      usersRepo.deleteById(userID);
+      return "User with id " + userID + " deleted.";
+    } else {
+      throw new UserNotFoundException(userID);
+    }
   }
 
   @Override
   public UsersDTO getUserByID(Long userID) {
-    return convertToUserDTO(usersRepo.getOne(userID));
+    Users user;
+    if (usersRepo.findById(userID).isPresent()) {
+      user = usersRepo.getOne(userID);
+      return convertToUserDTO(user);
+    } else {
+      throw new UserNotFoundException(userID);
+    }
   }
 
   @Override
   public UsersDTO updateUser(UsersDTO user) {
-    Users userToUpdate = null;
     if (usersRepo.findById(user.getId()).isPresent()) {
-      userToUpdate = usersRepo.findById(user.getId()).get();
+      Users userToUpdate = usersRepo.findById(user.getId()).get();
       userToUpdate.setEmail(user.getEmail());
       userToUpdate.setUserName(user.getUserName());
+      usersRepo.save(userToUpdate);
       return convertToUserDTO(userToUpdate);
+    } else {
+      throw new UserNotFoundException(user.getId());
     }
-    return user;
   }
 
   @Override
-  public String changePassword(String oldPassword, String newPassword) {
-    return null;
+  public String changePassword(Long userId, String oldPassword, String newPassword) {
+    if (usersRepo.findById(userId).isPresent()) {
+      if (usersRepo.findById(userId).get().getPassword().equals(oldPassword)) {
+        Users userToUpdate = usersRepo.findById(userId).get();
+        userToUpdate.setPassword(newPassword);
+        usersRepo.save(userToUpdate);
+        return "Password updated";
+      }
+      return "The provided password is not matching the old";
+    } else {
+      throw new UserNotFoundException(userId);
+    }
   }
 
   @Override
-  public boolean checkIfUserNameExists(String userName) {
-    return false;
+  public boolean checkUserNameExists(String userName) {
+    List<Users> u = usersRepo.findByUserName(userName);
+    return !u.isEmpty();
   }
 
   @Override
-  public boolean checkIfEmailExists(String email) {
-    return false;
+  public boolean checkEmailExists(String email) {
+    List<Users> u = usersRepo.findByEmail(email);
+    return !u.isEmpty();
   }
 
   private UsersDTO convertToUserDTO(Users user) {
